use integer for now

+++ integration_tests/modules_math.py
@@ -1,7 +1,7 @@
-def factorial(x: i32) -> i64:
+def factorial(x: i32) -> i32:
     if x < 0:
         return 0
-    result: i64
+    result: i32
     result = 1
     while x > 0:
         result = result * x
